💥 Change repo.isReleasedVersion() and repo.isOlderReleasedVersion() functions to filters
Nunjucks does not support asynchronous functions.
However, it does support asynchronous filters.
